project/requirements.txt
asgiref==3.8.1
channels==3.0.4
daphne==3.0.2
Django==4.2.1
django-allauth==0.63.3
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
simplejson==3.19.2
Pillow==9.0.0
